breaking the app into small pieces to avoid monolithic app
