+++ .idea/jarRepositories.xml
@@ -16,5 +16,10 @@
       <option name="name" value="JBoss Community repository" />
       <option name="url" value="https://repository.jboss.org/nexus/content/repositories/public/" />
     </remote-repository>
+    <remote-repository>
+      <option name="id" value="gpsUtil" />
+      <option name="name" value="gpsUtil" />
+      <option name="url" value="file://C:\Users\B294LQ\OneDrive - AXA\Desktop\Formation Java\Projet 8\P8-TourGuide-PM\TourGuide\libs\gpsUtil.jar" />
+    </remote-repository>
   </component>
 </project>
