+++ TourGuide/src/test/java/com/phildev/tourguide/TestUserService.java
@@ -1,0 +1,90 @@
+package com.phildev.tourguide;
+
+import com.phildev.tourguide.helper.InternalTestHelper;
+import com.phildev.tourguide.service.RewardsService;
+import com.phildev.tourguide.service.UserService;
+import com.phildev.tourguide.user.User;
+import gpsUtil.GpsUtil;
+import gpsUtil.location.VisitedLocation;
+import org.junit.jupiter.api.Test;
+import rewardCentral.RewardCentral;
+
+import java.util.List;
+import java.util.UUID;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+public class TestUserService {
+
+    @Test
+    public void getUserLocation() {
+        GpsUtil gpsUtil = new GpsUtil();
+        RewardsService rewardsService = new RewardsService(gpsUtil, new RewardCentral());
+        InternalTestHelper.setInternalUserNumber(0);
+        UserService userService = new UserService(gpsUtil, rewardsService);
+
+        User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
+        VisitedLocation visitedLocation = userService.trackUserLocation(user);
+        userService.tracker.stopTracking();
+        assertEquals(visitedLocation.userId, user.getUserId());
+    }
+
+    @Test
+    public void addUser() {
+        GpsUtil gpsUtil = new GpsUtil();
+        RewardsService rewardsService = new RewardsService(gpsUtil, new RewardCentral());
+        InternalTestHelper.setInternalUserNumber(0);
+        UserService userService = new UserService(gpsUtil, rewardsService);
+
+        User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
+        User user2 = new User(UUID.randomUUID(), "jon2", "000", "jon2@tourGuide.com");
+
+        userService.addUser(user);
+        userService.addUser(user2);
+
+        User retrievedUser = userService.getUser(user.getUserName());
+        User retrievedUser2 = userService.getUser(user2.getUserName());
+
+        userService.tracker.stopTracking();
+
+        assertEquals(user, retrievedUser);
+        assertEquals(user2, retrievedUser2);
+    }
+
+    @Test
+    public void getAllUsers() {
+        GpsUtil gpsUtil = new GpsUtil();
+        RewardsService rewardsService = new RewardsService(gpsUtil, new RewardCentral());
+        InternalTestHelper.setInternalUserNumber(0);
+        UserService userService = new UserService(gpsUtil, rewardsService);
+
+        User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
+        User user2 = new User(UUID.randomUUID(), "jon2", "000", "jon2@tourGuide.com");
+
+        userService.addUser(user);
+        userService.addUser(user2);
+
+        List<User> allUsers = userService.getAllUsers();
+
+        userService.tracker.stopTracking();
+
+        assertTrue(allUsers.contains(user));
+        assertTrue(allUsers.contains(user2));
+    }
+
+    @Test
+    public void trackUser() {
+        GpsUtil gpsUtil = new GpsUtil();
+        RewardsService rewardsService = new RewardsService(gpsUtil, new RewardCentral());
+        InternalTestHelper.setInternalUserNumber(0);
+        UserService userService = new UserService(gpsUtil, rewardsService);
+
+        User user = new User(UUID.randomUUID(), "jon", "000", "jon@tourGuide.com");
+        VisitedLocation visitedLocation = userService.trackUserLocation(user);
+
+        userService.tracker.stopTracking();
+
+        assertEquals(user.getUserId(), visitedLocation.userId);
+    }
+}
